Add HighlightTag::Operator, use it for unsafe deref. Move unsafe validation to its own file
